requirements.txt
aiohttp==3.7.3
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
decorator==4.4.2
future==0.18.2
geocoder==1.38.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
progressbar==2.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
ratelim==0.1.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
South==1.0.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
yarl==1.6.3
